format script and run it before commit
