+++ .eslintrc.json
@@ -18,6 +18,7 @@
   "plugins": ["react", "prettier"],
   "rules": {
     "quotes": ["error", "single"],
-    "indent": ["error", 2]
+    "indent": ["error", 2],
+    "jsx-quotes": ["error", "prefer-single"]
   }
 }
